% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2graph.R
\name{df_to_graph}
\alias{df_to_graph}
\title{Transforms a Data Frame to an igraph graph object.}
\usage{
df_to_graph(df)
}
\arguments{
\item{df}{A data frame. Each cell of the data frame will become a node, so be careful
with the chosen data frame... Long contents might not be the best idea...}
}
\value{
An (directed) igraph graph object, that can be plotted... Duplicate edges are
removed.
}
\description{
Transforms a Data Frame to an igraph graph object.
}
\examples{
df <- data.frame(a = c(1,2,3), b = c(1,3,4))
g <- df_to_graph(df)
plot(g)
}
