% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2graph.R
\name{.df_to_graph_edges_list}
\alias{.df_to_graph_edges_list}
\title{Create a complete edges list from a Data Frame.}
\usage{
.df_to_graph_edges_list(df)
}
\arguments{
\item{df}{A data frame for which each cell will become a node, and each row in turn will
be transformed into a list of "edges" for a future graph}
}
\value{
A list of edges, in matrix form.
}
\description{
Create a complete edges list from a Data Frame.
}
