% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2graph.R
\name{.row_to_graph_edge_list}
\alias{.row_to_graph_edge_list}
\title{Transform DF row to an edges list.}
\usage{
.row_to_graph_edge_list(x)
}
\arguments{
\item{x}{A dataframe's row.}
}
\value{
A matrix of two columns, whereby each row of the matrix is a future graph
edge.
}
\description{
Transform DF row to an edges list.
}
